Rename files containing classes.
