+++ lib/CborldError.js
