+++ lib/codec.js
@@ -1,14 +1,14 @@
 /*!
  * Copyright (c) 2020 Digital Bazaar, Inc. All rights reserved.
  */
-import {Base64PadCodec} from './codecs/base64Pad';
-import {CborldError} from './error';
-import {CodecMapCodec} from './codecs/codecMap';
-import {ContextCodec} from './codecs/context';
-import {SimpleTypeCodec} from './codecs/simpleType';
-import {UrlCodec} from './codecs/url';
-import {XsdDateTimeCodec} from './codecs/xsdDateTime';
-import {termCodecs} from './codecs/codecMap';
+import {Base64PadCodec} from './codecs/Base64PadCodec';
+import {CborldError} from './CborldError';
+import {CodecMapCodec} from './codecs/CodecMapCodec';
+import {ContextCodec} from './codecs/ContextCodec';
+import {SimpleTypeCodec} from './codecs/SimpleTypeCodec';
+import {UrlCodec} from './codecs/UrlCodec';
+import {XsdDateTimeCodec} from './codecs/XsdDateTimeCodec';
+import {termCodecs} from './codecs/CodecMapCodec';
 
 /**
  * Gets a list of all term codecs available to use when transforming term
