+++ lib/codecs/Base64PadCodec.js
