+++ lib/codecs/CodecMapCodec.js
@@ -2,9 +2,9 @@
  * Copyright (c) 2020 Digital Bazaar, Inc. All rights reserved.
  */
 // map of future CBOR-LD codec registry of known CBOR-LD codecs
-import {Base64PadCodec} from './base64Pad';
-import {UrlCodec} from './url';
-import {XsdDateTimeCodec} from './xsdDateTime';
+import {Base64PadCodec} from './Base64PadCodec';
+import {UrlCodec} from './UrlCodec';
+import {XsdDateTimeCodec} from './XsdDateTimeCodec';
 
 const codecEncodeMap = new Map();
 const codecDecodeMap = new Map();
