+++ lib/codecs/ContextCodec.js
@@ -1,7 +1,7 @@
 /*!
  * Copyright (c) 2020 Digital Bazaar, Inc. All rights reserved.
  */
-import {CborldError} from '../error';
+import {CborldError} from '../CborldError';
 
 // map of future CBOR-LD registry of known JSON-LD Contexts
 const cborldContextEncodeMap = new Map();
