+++ lib/codecs/SimpleTypeCodec.js
