+++ lib/codecs/UncompressedCborldCodec.js
