+++ lib/context.js
@@ -1,8 +1,8 @@
 /*!
  * Copyright (c) 2020 Digital Bazaar, Inc. All rights reserved.
  */
-import {CborldError} from './error';
-import {ContextCodec} from './codecs/context';
+import {CborldError} from './CborldError';
+import {ContextCodec} from './codecs/ContextCodec';
 
 /**
 * Gets a list of JSON-LD Contexts given a JSON-LD Document.
