+++ lib/decode.js
@@ -1,7 +1,7 @@
 /*!
  * Copyright (c) 2020 Digital Bazaar, Inc. All rights reserved.
  */
-import {CborldError} from './error';
+import {CborldError} from './CborldError';
 import cbor from 'cbor';
 import {getCborldContextUrls} from './context';
 import {getTermCodecMap} from './codec';
