+++ lib/encode.js
@@ -1,10 +1,10 @@
 /*!
  * Copyright (c) 2020 Digital Bazaar, Inc. All rights reserved.
  */
-import {CborldError} from './error';
-import {CodecMapCodec} from './codecs/codecMap';
-import {CompressedCborldCodec} from './codecs/compressedCborld';
-import {UncompressedCborldCodec} from './codecs/uncompressedCborld';
+import {CborldError} from './CborldError';
+import {CodecMapCodec} from './codecs/CodecMapCodec';
+import {CompressedCborldCodec} from './codecs/CompressedCborldCodec';
+import {UncompressedCborldCodec} from './codecs/UncompressedCborldCodec';
 import cbor from 'cbor';
 import {getJsonldContextUrls} from './context';
 import {getTermCodecMap} from './codec';
@@ -26,8 +26,9 @@ import {getTermCodecMap} from './codec';
  *
  * @returns {Promise<Uint8Array>} - The encoded CBOR-LD bytes.
  */
-export async function toCborld({
-  jsonldDocument, documentLoader, appContextMap, appTermMap = new Map(), diagnose}) {
+export async function toCborld(
+  {jsonldDocument, documentLoader, appContextMap, appTermMap = new Map(),
+    diagnose}) {
 
   let termEncodingMap;
